OcdFileExport: Factor out icon support
The new class OcdIcon allows to handle icon export in a
functional style, in a separate, encapsulated unit.
On this occassion, adjust some types and identifiers for better
consistency.
