+++ src/fileformats/ocd_file_export.cpp
@@ -23,9 +23,7 @@
 #include "ocd_file_export.h"
 
 #include <algorithm>
-#include <array>
 #include <cstddef>
-#include <cstdlib>
 #include <iterator>
 #include <memory>
 #include <utility>
@@ -35,19 +33,16 @@
 #include <Qt>
 #include <QtGlobal>
 #include <QtMath>
-#include <QColor>
 #include <QCoreApplication>
 #include <QFileInfo>
 #include <QFlags>
 #include <QFontMetricsF>
 #include <QIODevice>
-#include <QImage>
 #include <QLatin1Char>
 #include <QLatin1String>
 #include <QPoint>
 #include <QPointF>
 #include <QRectF>
-#include <QRgb>
 #include <QString>
 #include <QTextCodec>
 #include <QTextDecoder>
@@ -77,6 +72,7 @@
 #include "fileformats/file_import_export.h"
 #include "fileformats/ocd_file_format.h"
 #include "fileformats/ocd_georef_fields.h"
+#include "fileformats/ocd_icon.h"
 #include "fileformats/ocd_types.h"
 #include "fileformats/ocd_types_v8.h"
 #include "fileformats/ocd_types_v9.h"
@@ -291,115 +287,6 @@ constexpr qint32 convertSize(qint64 size)
 int convertRotation(qreal angle)
 {
 	return qRound(10 * qRadiansToDegrees(angle));
-}
-
-
-int getPaletteColorV6(QRgb rgb)
-{
-	Q_ASSERT(qAlpha(rgb) == 255);
-	
-	// Quickly return for most frequent value
-	if (rgb == qRgb(255, 255, 255))
-		return 15;
-	
-	auto color = QColor(rgb).toHsv();
-	if (color.hue() == -1 || color.saturation() < 32)
-	{
-		auto gray = qGray(rgb);  // qGray is used for dithering
-		if (gray >= 192)
-			return 8;
-		if (gray >= 128)
-			return 7;
-		return 0;
-	}
-	
-	struct PaletteColor
-	{
-		int hue;
-		int saturation;
-		int value;
-		static PaletteColor fromQColor(const QColor& color)
-		{
-			return { color.hue(), color.saturation(), color.value() };
-		}
-	};
-	static const auto palette = []() {
-		std::array<PaletteColor, 16> palette_hsv;
-		const auto palette_rgb = Ocd::IconV8::palette<QColor>();
-		std::transform(begin(palette_rgb), end(palette_rgb), begin(palette_hsv), [](auto& rgb){
-			return PaletteColor::fromQColor(rgb.toHsv());
-		});
-		return palette_hsv;
-	}();
-	
-	int best_index = 0;
-	auto best_distance = 2100000;  // > 6 * (10*sq(180) + sq(128) + sq(64))
-	auto sq = [](int n) { return n*n; };
-	for (auto i : { 1, 2, 3, 4, 5, 6, 9, 10, 11, 12, 13, 14 })
-	{
-		// True color
-		const auto& palette_color = palette[i];
-		auto hue_dist = std::abs(color.hue() - palette_color.hue);
-		auto distance = 10 * sq(std::min(hue_dist, 360 - hue_dist))
-		                + sq(color.saturation() - palette_color.saturation)
-		                + sq(color.value() - palette_color.value);
-		
-		// (Too much) manual tweaking for orienteering colors
-		if (i == 1)
-			distance *= 3;	// Dark red
-		else if (i == 3)
-			distance *= 4;		// Olive
-		else if (i == 11)
-			distance *= 4;		// Yellow
-		else if (i == 9)
-			distance *= 6;		// Red is unlikely
-		else
-			distance *= 2;
-		
-		if (distance < best_distance)
-		{
-			best_distance = distance;
-			best_index = i;
-		}
-	}
-	return best_index;
-}
-
-
-quint8 getPaletteColorV9(QRgb rgb)
-{
-	Q_ASSERT(qAlpha(rgb) == 255);
-	
-	// Quickly return most frequent value
-	if (rgb == qRgb(255, 255, 255))
-		return 124;
-	
-	struct PaletteColor
-	{
-		int r;
-		int g;
-		int b;
-	};
-	static const auto palette = Ocd::IconV9::palette<PaletteColor>();
-	
-	auto r = qRed(rgb);
-	auto g = qGreen(rgb);
-	auto b = qBlue(rgb);
-	auto sq = [](int n) { return n*n; };
-
-	quint8 best_index = 0;
-	auto best_distance = 10000; // > (2 + 3 + 4) * sq(32)
-	for (quint8 i = 0; i < 125; ++i)
-	{
-		auto palette_color = palette[i];
-		auto distance = 2 * sq(r - palette_color.r) + 4 * sq(g - palette_color.g) + 3 * sq(b - palette_color.b);
-		if (distance < best_distance)
-		{
-			best_distance = distance;
-			best_index = i;
-		}
-	}
-	return best_index;
 }
 
 
@@ -1180,7 +1067,7 @@ void OcdFileExport::setupBaseSymbol(const Symbol* symbol, quint32 symbol_number,
 		}
 	}
 	
-	exportSymbolIcon(symbol, ocd_base_symbol.icon);
+	ocd_base_symbol.icon = OcdIcon{*map, *symbol};
 }
 
 
@@ -2129,7 +2016,7 @@ QByteArray OcdFileExport::exportCombinedAreaSymbol(
 {
 	auto ocd_symbol = exportAreaSymbol<OcdAreaSymbol>(area_symbol, symbol_number);
 	auto ocd_subsymbol_data = reinterpret_cast<OcdAreaSymbol*>(ocd_symbol.data());
-	exportSymbolIcon(combined_symbol, ocd_subsymbol_data->base.icon);
+	ocd_subsymbol_data->base.icon = OcdIcon{*map, *combined_symbol};
 	ocd_subsymbol_data->common.border_on_V9 = 1;
 	ocd_subsymbol_data->border_symbol = symbol_numbers[line_symbol];
 	return ocd_symbol;
@@ -2146,7 +2033,7 @@ QByteArray OcdFileExport::exportCombinedLineSymbol(
 {
 	auto ocd_symbol = exportLineSymbol<OcdLineSymbol>(main_line, symbol_number);
 	auto ocd_symbol_data = reinterpret_cast<OcdLineSymbol*>(ocd_symbol.data());
-	exportSymbolIcon(combined_symbol, ocd_symbol_data->base.icon);
+	ocd_symbol_data->base.icon = OcdIcon{*map, *combined_symbol};
 	
 	auto& ocd_line_common = ocd_symbol_data->common;
 	if (framing)
@@ -2177,89 +2064,6 @@ QByteArray OcdFileExport::exportCombinedLineSymbol(
 	}
 	
 	return ocd_symbol;
-}
-
-
-
-void OcdFileExport::exportSymbolIcon(const Symbol* symbol, Ocd::IconV8& icon)
-{
-	// Icon: 22x22 with 4 bit palette color, origin at bottom left
-	QImage image = symbol->createIcon(*map, std::max(icon.width(), icon.height()), false)
-	               .convertToFormat(QImage::Format_ARGB32_Premultiplied);
-	
-	auto process_pixel = [&image](int x, int y)->int {
-		// Apply premultiplied pixel on white background
-		auto premultiplied = image.pixel(x, y);
-		auto alpha = qAlpha(premultiplied);
-		auto r = 255 - alpha + qRed(premultiplied);
-		auto g = 255 - alpha + qGreen(premultiplied);
-		auto b = 255 - alpha + qBlue(premultiplied);
-		auto pixel = qRgb(r, g, b);
-		
-		// Ordered dithering 2x2 threshold matrix, adjusted for o-map halftones
-		static int threshold[4] = { 24, 192, 136, 80 };
-		auto palette_color = getPaletteColorV6(pixel);
-		switch (palette_color)
-		{
-		case 0:
-			// Black to gray (50%)
-			return  qGray(pixel) < 128-threshold[(x%2 + 2*(y%2))]/2 ? 0 : 7;
-			
-		case 7:
-			// Gray (50%) to light gray 
-			return  qGray(pixel) < 192-threshold[(x%2 + 2*(y%2))]/4 ? 7 : 8;
-			
-		case 8:
-			// Light gray to white
-			return  qGray(pixel) < 256-threshold[(x%2 + 2*(y%2))]/4 ? 8 : 15;
-			
-		case 15:
-			// Pure white
-			return palette_color;
-			
-		default:
-			// Color to white
-			return  QColor(pixel).saturation() >= threshold[(x%2 + 2*(y%2))] ? palette_color : 15;
-		}
-	};
-	
-	auto icon_bits = icon.bits;
-	for (int y = icon.height() - 1; y >= 0; --y)
-	{
-		for (int x = 0; x < icon.width(); x += 2)
-		{
-			auto first = process_pixel(x, y);
-			auto second = process_pixel(x+1, y);
-			*(icon_bits++) = quint8((first << 4) + second);
-		}
-		icon_bits++;
-	}
-}
-
-void OcdFileExport::exportSymbolIcon(const Symbol* symbol, Ocd::IconV9& icon)
-{
-	// Icon: 22x22 with 8 bit palette color code, origin at bottom left
-	QImage image = symbol->createIcon(*map, std::max(icon.width(), icon.height()), true)
-	               .convertToFormat(QImage::Format_ARGB32_Premultiplied);
-	
-	auto process_pixel = [&image](int x, int y)->quint8 {
-		// Apply premultiplied pixel on white background
-		auto premultiplied = image.pixel(x, y);
-		auto alpha = qAlpha(premultiplied);
-		auto r = 255 - alpha + qRed(premultiplied);
-		auto g = 255 - alpha + qGreen(premultiplied);
-		auto b = 255 - alpha + qBlue(premultiplied);
-		return getPaletteColorV9(qRgb(r, g, b));
-	};
-	
-	auto icon_bits = icon.bits;
-	for (int y = icon.height() - 1; y >= 0; --y)
-	{
-		for (int x = 0; x < icon.width(); ++x)
-		{
-			*(icon_bits++) = process_pixel(x, y);
-		}
-	}
 }
